move dynamicdynamicvariables to a new location
